change int to float

+++ database/migrations/2019_05_22_171726_create_wallet_accounts.php
@@ -20,7 +20,7 @@ class CreateWalletAccounts extends Migration
             $table->string('color');
             $table->boolean('exclude_from_stats');
             $table->boolean('gps');
-            $table->integer('init_amount');
+            $table->float('init_amount');
             $table->float('balance')->default(0);
             $table->integer('position');
             $table->timestamps();
